test flush with MeterProviderInterface
